Add getCenterCood and Completely split registerEvents
